Move `SpyCollectionViewController` to own file
